Move common integration tests based functionality into a base class and update existing integration tests to inherit from it.
