+++ st2api/tests/integration/test_gunicorn_configs.py
@@ -24,13 +24,14 @@ from eventlet.green import subprocess
 
 from st2common.models.utils import profiling
 from st2common.util.shell import kill_process
+from st2tests.base import IntegrationTestCase
 
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 ST2_CONFIG_PATH = os.path.join(BASE_DIR, '../../../conf/st2.tests.conf')
 
 
-class GunicornWSGIEntryPointTestCase(unittest2.TestCase):
+class GunicornWSGIEntryPointTestCase(IntegrationTestCase):
     @unittest2.skipIf(profiling.is_enabled(), 'Profiling is enabled')
     def test_st2api_wsgi_entry_point(self):
         port = random.randint(10000, 30000)
@@ -40,7 +41,7 @@ class GunicornWSGIEntryPointTestCase(unittest2.TestCase):
         env['ST2_CONFIG_PATH'] = ST2_CONFIG_PATH
         process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env,
                                    shell=True, preexec_fn=os.setsid)
-
+        self.add_process(process=process)
         eventlet.sleep(5)
         response = requests.get('http://127.0.0.1:%s/v1/actions' % (port))
         self.assertEqual(response.status_code, httplib.OK)
@@ -56,6 +57,7 @@ class GunicornWSGIEntryPointTestCase(unittest2.TestCase):
         process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env,
                                    shell=True, preexec_fn=os.setsid)
 
+        self.add_process(process=process)
         eventlet.sleep(5)
         response = requests.post('http://127.0.0.1:%s/tokens' % (port))
         self.assertEqual(response.status_code, httplib.UNAUTHORIZED)
